refactor: gui.widgets package
