+++ src/viperleed/gui/leedsim/dialogs/dialogbulk3dsym.py
@@ -18,7 +18,7 @@ import PyQt5.QtGui as qtg
 import PyQt5.QtWidgets as qtw
 
 from viperleed.gui.classes import planegroup
-from viperleed.gui.widgetslib import AllGUIFonts
+from viperleed.gui.widgets.lib import AllGUIFonts
 
 PlaneGroup = planegroup.PlaneGroup
