+++ src/viperleed/gui/leedsim/widgets/editablematrix.py
@@ -16,7 +16,7 @@ import PyQt5.QtCore as qtc
 import PyQt5.QtGui as qtg
 import PyQt5.QtWidgets as qtw
 
-from viperleed.gui.widgetslib import AllGUIFonts
+from viperleed.gui.widgets.lib import AllGUIFonts
 
 from viperleed.gui import decorators as dev_
