+++ src/viperleed/gui/widgets/__init__.py
@@ -1,0 +1,31 @@
+"""Package widgets of viperleed.gui.
+
+Modules
+-------
+basewidgets
+    Generic widgets that do not fit other modules of the widgets
+    packages.
+buttons
+    Custom buttons and button-related widgets.
+canvases
+    Subclasses of matplotlib's FigureCanvas.
+checkboxes
+    Custom check boxes and radio buttons
+collapsible
+    Widgets that can be expanded/collapsed with user clicks.
+decorators
+    Decorators dedicated to Qt QWidgets.
+lib
+    Functions and classes for widgets and other Qt objects.
+spinboxes
+    Spin-box widgets (i.e., inputs of numerical values with
+    incremental steps).
+"""
+
+__authors__ = (
+    'Michele Riva (@michele-riva)',
+    'Florian Dörr (@FlorianDoerr)',
+    )
+__copyright__ = 'Copyright (c) 2019-2025 ViPErLEED developers'
+__created__ = '2024-06-13'
+__license__ = 'GPLv3+'
